Finish testing UserController level API with Postman

+++ src/main/java/com/sprint/mission/discodeit/dto/binarycontentdto/BinaryContentDTO.java
@@ -7,7 +7,7 @@ public record BinaryContentDTO(
     UUID id,
     Instant createdAt,
     String fileName,
-    int size,
+    long size,
     String contentType,
     byte[] bytes
 ) {
